Made two new windows, added buttons to second window, implemented heatmap from file data
Added two buttons on the second window. On of them opens the colocalization window, other opens the biological process analyzer window. Implemented colocalization window and biological process analyzer window. Added heatmap from Adrian into both, changed it to use data from the file in the colocalization window, not 100% sure that it is correct because of indexing. Heatmap appears on colocalization window with the file data. Changed recData.csv to allow for heatmaps.