Added functionalities to bioprocess window, small improvements to colocalization window, minor style adjustments
Bioprocess window now has an up to date heatmap. The heatmap can be saved.
Added a text edit for the user to input the biological process to analyse. Included error handling.
Added a method where we will have a vector with all the processes we can analyse, only added hypoxia for now.
Added comments and deleted obsolete code in colocalization window.
Changed border colour back to black on second window.