Implemented biological process 2
Implemented the computation of the Sliced Earth Mover Distance. Implemented the k-Medoid clustering of genes using the EMD matrix, which returns a vector of strings to pass to the API team. However, there might still be some issues with these functions, especially if we do not do simple filtering beforehand.

Co-Authored-By: Lorenzo Tarricone <96067929+LorenzoTarricone@users.noreply.github.com>