Attempting to solve issues in api_gene_name.cpp and colocalisation
There were lots of issues when trying to delete variables: in api_gene_name the set destructors seemed to crash the program. In colocalisation, the class destructors deletes the pointers to the matrices which are stored on the heap, however we already deleted these pointers in the intermediate steps of the colocalisation matrix, so trying to delete them again was causing the program to crash. Also fixed an issue where we were removing some characters from gene names during the api process, now we should only remove blank spaces in the string